made int result type functions require int parameters

+++ virtdata-lib-basics/src/main/java/io/virtdata/libbasics/shared/from_long/to_int/Add.java
@@ -9,7 +9,7 @@ public class Add implements LongToIntFunction {
 
     private final long addend;
 
-    public Add(long addend) {
+    public Add(int addend) {
         this.addend = addend;
     }
